Upload monitorings from db and generate CSV only when export to zip
