+++ app/src/main/kotlin/org/bspb/smartbirds/pro/service/SyncService.kt
@@ -50,7 +50,7 @@ open class SyncService : AbstractIntentService("SyncService") {
             try {
                 isWorking = true
                 updateSyncProgress(R.string.upload_dialog_text)
-                uploadManager.uploadAllNew()
+                uploadManager.uploadAll()
                 fetchNewData()
             } finally {
                 isWorking = false
