+++ app/src/main/kotlin/org/bspb/smartbirds/pro/ui/fragment/BrowseMonitoringCommonFormFragment.kt
@@ -10,7 +10,6 @@ import org.bspb.smartbirds.pro.R
 import org.bspb.smartbirds.pro.SmartBirdsApplication
 import org.bspb.smartbirds.pro.content.Monitoring
 import org.bspb.smartbirds.pro.events.EEventBus
-import org.bspb.smartbirds.pro.service.DataOpsService_
 import org.bspb.smartbirds.pro.ui.utils.Configuration
 import org.bspb.smartbirds.pro.utils.MonitoringManager
 
@@ -48,7 +47,6 @@ open class BrowseMonitoringCommonFormFragment : BaseCommonFormFragment() {
                 commonForm.clear()
                 commonForm.putAll(data)
                 monitoringManager.update(this@run)
-                DataOpsService_.intent(context).generateMonitoringFiles(code).start()
             }
         }
     }
